Split rendering logic into separate file
This introduces a resume builder which can be used to make finer-grain modifications to a supplied resume JSON without requiring changes to the JSON itself. This way, the JSON resume can be representative of an entire career versus just a single iteration of a resume.

For a proof of concept, there is a single method on the builder which currently shows this pattern called `experiences`. This method takes an object allowing the user to specify the exact names of experiences to filter and a date after which to filter out.

Other improvements include:
- added an `indent` function to more cleanly specify the amount of indentation being done within the rendering process when necessary
- added sorting of experiences by date to ensure descending ordering
- added checks to not render anything for skills, experience, or projects if there are none, previously this would cause an error because the section was still rendered