Support local configs
We'll first walk through all the ancestors of the current working
directory (including the current working directory) and look for the
nearest file called "reze.toml".

If no such file is found, we'll check the global config path,
".config/reze/reze.toml", for a config.

If neither is found we give up.

Currently we do not check that file is well-formed when walking up the
CWD ancestors, so in the situation that the nearest config is malformed
but there is another farther well-formed config, we won't return that
farther config.

Signed-off-by: Andrew Leung <andrewleung104@gmail.com>